restructuring wake model files
