+++ .github/workflows/python-package.yml
@@ -25,10 +25,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest
-        #if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        #export PYTHONPATH=`pwd`:$PYTHONPATH
-        python -m pip install .
+        python -m pip install .[test]
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
