+++ foxes/models/partial_wakes/mapped.py
@@ -5,15 +5,9 @@ from foxes.models.partial_wakes.rotor_points import RotorPoints
 from foxes.models.partial_wakes.top_hat import PartialTopHat
 from foxes.models.partial_wakes.axiwake import PartialAxiwake
 from foxes.models.partial_wakes.distsliced import PartialDistSlicedWake
-from foxes.models.wake_models.dist_sliced.axisymmetric.top_hat.top_hat_wake_model import (
-    TopHatWakeModel,
-)
-from foxes.models.wake_models.dist_sliced.dist_sliced_wake_model import (
-    DistSlicedWakeModel,
-)
-from foxes.models.wake_models.dist_sliced.axisymmetric.axisymmetric_wake_model import (
-    AxisymmetricWakeModel,
-)
+from foxes.models.wake_models.top_hat import TopHatWakeModel
+from foxes.models.wake_models.dist_sliced import DistSlicedWakeModel
+from foxes.models.wake_models.axisymmetric import AxisymmetricWakeModel
 
 
 class Mapped(PartialWakesModel):
