--- foxes/models/wake_models/dist_sliced/__init__.py
@@ -1,3 +1,0 @@
-from .dist_sliced_wake_model import DistSlicedWakeModel
-
-from . import axisymmetric
