--- foxes/models/wake_models/dist_sliced/axisymmetric/__init__.py
@@ -1,4 +1,0 @@
-from .axisymmetric_wake_model import AxisymmetricWakeModel
-
-from . import gaussian
-from . import top_hat
