--- foxes/models/wake_models/dist_sliced/axisymmetric/gaussian/__init__.py
@@ -1,2 +1,0 @@
-from .gaussian_wake_model import GaussianWakeModel
-from .bastankhah import BastankhahWake
