+++ foxes/models/wake_models/ti/__init__.py
@@ -1,2 +1,1 @@
-from .jensen import JensenWake
 from .crespo_hernandez import CrespoHernandezTIWake
