+++ setup.cfg
@@ -39,6 +39,7 @@ install_requires =
 
 [options.extras_require]
 test =
+    flake8
     pytest
 doc = 
     sphinx
